Add patch for AMD ZEN to fix microphone issue
In the function AppleHDAController::getAudioStreamLinkPositionInDMABuffer(IOHDAStream*, uint64_t*),
it expects bit 0 of register 0x70 to be cleared before reading DMA position.
However, for AMD, it somehow is not cleared, hence DMA position is incorrect.
For now the workround is to skip that bit check.

This patch has been veried to work up to macOS Sonoma.